Features
Added Date parser
Many permutations are permitted with different splitting characters.

Running flag
If the drop function is called the Running flag is set to true.
If the CheckGateDirectory function is called the Running flag is set to true.
The timer nor the drop function cannot be executed when the Running flag is set to true
Once the OrganiceFiles function finishes is sets the Running flag to false.